Change the polls pagination parameters
